+++ src/commands/index.ts
@@ -1,0 +1,3 @@
+export * from './convert-text';
+export * from './set-base-path';
+export * from './set-prefix';