+++ src/enums/storage-key.ts
@@ -1,0 +1,11 @@
+export enum StorageKey {
+  basePath = 'translationHelper.basePath',
+  prefix = 'translationHelper.prefix',
+  maxWords = 'maxWords',
+  translationCodeMask = 'translationCodeMask',
+  caseType = 'caseType',
+  autoEditTranslationFiles = 'autoEditTranslationFiles',
+  translationsPath = 'translationsPath',
+  supportedLanguages = 'supportedLanguages',
+  translationKeyStyle = 'translationKeyStyle',
+}