+++ src/enums/translation-key-style.ts
@@ -1,0 +1,4 @@
+export enum TranslationKeyStyle {
+  nested = 'nested',
+  flat = 'flat'
+}