+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## Angular Translation Helper Changelog
 
+<a name="0.0.2"></a>
+
+# 0.0.2 (2022-06-13)
+
+- Add translation files support
+- Add nested/flat translation key styles support 
+
 <a name="0.0.1"></a>
 
 # 0.0.1 (2022-06-10)